+++ sdk/resourcemanager/elastic/armelastic/CHANGELOG.md
@@ -1,5 +1,49 @@
 # Release History
 
+## 0.6.0 (2022-06-05)
+### Features Added
+
+- New const `APIVersionParameterTwoThousandTwentyOne0901Preview`
+- New const `APIVersionParameterTwoThousandTwentyTwo0505Preview`
+- New const `APIVersionParameterTwoThousandTwenty0701Preview`
+- New const `TypeIP`
+- New const `APIVersionParameterTwoThousandTwenty0701`
+- New const `APIVersionParameterTwoThousandTwentyOne1001Preview`
+- New const `TypeAzurePrivateEndpoint`
+- New function `PossibleAPIVersionParameterValues() []APIVersionParameter`
+- New function `PossibleTypeValues() []Type`
+- New struct `AllTrafficFiltersClientListOptions`
+- New struct `AllTrafficFiltersClientListResponse`
+- New struct `AssociateTrafficFilterClientAssociateResponse`
+- New struct `AssociateTrafficFilterClientBeginAssociateOptions`
+- New struct `CreateAndAssociateIPFilterClientBeginCreateOptions`
+- New struct `CreateAndAssociateIPFilterClientCreateResponse`
+- New struct `CreateAndAssociatePLFilterClientBeginCreateOptions`
+- New struct `CreateAndAssociatePLFilterClientCreateResponse`
+- New struct `DetachAndDeleteTrafficFilterClientBeginDeleteOptions`
+- New struct `DetachAndDeleteTrafficFilterClientDeleteResponse`
+- New struct `DetachTrafficFilterClientBeginUpdateOptions`
+- New struct `DetachTrafficFilterClientUpdateResponse`
+- New struct `ExternalUserClientCreateOrUpdateOptions`
+- New struct `ExternalUserClientCreateOrUpdateResponse`
+- New struct `ExternalUserCreationResponse`
+- New struct `ExternalUserInfo`
+- New struct `ListAssociatedTrafficFiltersClientListOptions`
+- New struct `ListAssociatedTrafficFiltersClientListResponse`
+- New struct `MonitorClientBeginUpgradeOptions`
+- New struct `MonitorClientUpgradeResponse`
+- New struct `MonitorUpgrade`
+- New struct `TrafficFilter`
+- New struct `TrafficFilterResponse`
+- New struct `TrafficFilterRule`
+- New struct `TrafficFiltersClientDeleteOptions`
+- New struct `TrafficFiltersClientDeleteResponse`
+- New struct `UpgradableVersionsClientDetailsOptions`
+- New struct `UpgradableVersionsClientDetailsResponse`
+- New struct `UpgradableVersionsList`
+- New field `Version` in struct `MonitorProperties`
+
+
 ## 0.5.0 (2022-05-17)
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/elastic/armelastic` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html) since version 0.5.0, which contains breaking changes.